% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{add_classification}
\alias{add_classification}
\title{Añadir clasificación a un data.frame tidy}
\usage{
add_classification(df, index_col = "index", value_col = "value")
}
\arguments{
\item{df}{data.frame con columnas de índice y valor.}

\item{index_col}{nombre de la columna que identifica el índice.}

\item{value_col}{nombre de la columna con el valor numérico del índice.}
}
\value{
data.frame con columnas phase3, magnitud y escenario.
}
\description{
Añadir clasificación a un data.frame tidy
}
