% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_icen_classification}
\alias{add_icen_classification}
\title{Añadir clasificación ICEN a un data.frame con columna value}
\usage{
add_icen_classification(df, value_col = "value")
}
\arguments{
\item{df}{data.frame con columnas date, value (y opcional index)}

\item{value_col}{nombre de la columna numérica con ICEN}
}
\value{
df con phase3, magnitud, escenario
}
\description{
Añadir clasificación ICEN a un data.frame con columna value
}
