% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{add_icen_event_classification}
\alias{add_icen_event_classification}
\title{Añadir clasificación ICEN por eventos (tablas) a un data.frame}
\usage{
add_icen_event_classification(df, year_col = "year", month_col = "month")
}
\arguments{
\item{df}{data.frame con columnas de año y mes}

\item{year_col}{nombre de columna año}

\item{month_col}{nombre de columna mes}
}
\value{
df con phase3_event, magnitud_event, escenario_event
}
\description{
Añadir clasificación ICEN por eventos (tablas) a un data.frame
}
