% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{bio_by_icen_phase}
\alias{bio_by_icen_phase}
\title{Resumen biológico por fase ICEN (eventos/tablas)}
\usage{
bio_by_icen_phase(icen_df, bio_df)
}
\arguments{
\item{icen_df}{data.frame con date, year, month, value}

\item{bio_df}{data.frame con date, value}
}
\value{
data.frame con medias y diferencias vs Neutro
}
\description{
Resumen biológico por fase ICEN (eventos/tablas)
}
