% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{icen_classify_event_vec}
\alias{icen_classify_event_vec}
\title{Clasificar muchos (year,month) por TABLAS de eventos (vectorizada)}
\usage{
icen_classify_event_vec(year, month)
}
\arguments{
\item{year}{vector int}

\item{month}{vector int (1..12)}
}
\value{
tibble phase3, magnitud, escenario (mismo largo)
}
\description{
Clasificar muchos (year,month) por TABLAS de eventos (vectorizada)
}
