% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{icen_classify_value}
\alias{icen_classify_value}
\title{Clasificar ICEN por valor (fase y magnitud)}
\usage{
icen_classify_value(x)
}
\arguments{
\item{x}{vector numérico ICEN}
}
\value{
tibble: phase3, magnitud, escenario
}
\description{
Clasificar ICEN por valor (fase y magnitud)
}
