% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{icen_event_tables}
\alias{icen_event_tables}
\title{Tablas de eventos ICEN (Niño/Niña) según cronología (pegada por el usuario)}
\usage{
icen_event_tables()
}
\value{
lista con tibble nino y nina
}
\description{
Tablas de eventos ICEN (Niño/Niña) según cronología (pegada por el usuario)
}
