% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{labcos_classify_value}
\alias{labcos_classify_value}
\title{Clasificar LABCOS por bins}
\usage{
labcos_classify_value(x)
}
\arguments{
\item{x}{valor numérico del índice.}
}
\value{
tibble con phase3, magnitud y escenario.
}
\description{
Clasificar LABCOS por bins
}
