% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lag_correlation}
\alias{lag_correlation}
\title{Correlación con rezagos entre índice y biomasa (u otra serie)}
\usage{
lag_correlation(index_df, bio_df, max_lag = 24)
}
\arguments{
\item{index_df}{data.frame con serie del índice.}

\item{bio_df}{data.frame con serie biológica (p.ej. biomasa).}

\item{max_lag}{máximo rezago (en unidades del periodo de la serie).}
}
\value{
data.frame/tibble con correlaciones por lag.
}
\description{
Correlación con rezagos entre índice y biomasa (u otra serie)
}
