% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{period_key}
\alias{period_key}
\title{Generar clave de periodo}
\usage{
period_key(date, period = c("month", "quarter", "semester", "year"))
}
\arguments{
\item{date}{vector Date/POSIXct.}

\item{period}{periodo (p.ej. "month", "year", "season3m").}
}
\value{
vector con clave del periodo.
}
\description{
Generar clave de periodo
}
