% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_index_bands_with_bio}
\alias{plot_index_bands_with_bio}
\title{Graficar bandas del índice junto con serie biológica}
\usage{
plot_index_bands_with_bio(
  index_df,
  bio_df,
  alpha_band = 0.18,
  index_label = NULL,
  bio_label = "Bio"
)
}
\arguments{
\item{index_df}{data.frame del índice.}

\item{bio_df}{data.frame de la serie biológica.}

\item{alpha_band}{transparencia de la banda (0-1).}

\item{index_label}{etiqueta para el índice.}

\item{bio_label}{etiqueta para la serie biológica.}
}
\value{
objeto ggplot.
}
\description{
Graficar bandas del índice junto con serie biológica
}
