% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_month.R
\name{season3m_center_month}
\alias{season3m_center_month}
\title{Mes central de una estación móvil de 3 meses}
\usage{
season3m_center_month(season3m)
}
\arguments{
\item{season3m}{etiqueta de estación 3m (según tu convención).}
}
\value{
entero 1..12 del mes central.
}
\description{
Mes central de una estación móvil de 3 meses
}
