% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{sign_classify_value}
\alias{sign_classify_value}
\title{Clasificar por signo (Positivo/Negativo/Neutro)}
\usage{
sign_classify_value(x, eps = 0)
}
\arguments{
\item{x}{valor numérico.}

\item{eps}{tolerancia para declarar neutro.}
}
\value{
tibble con phase3, magnitud y escenario.
}
\description{
Clasificar por signo (Positivo/Negativo/Neutro)
}
