% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_month.R
\name{wide3m_to_monthly}
\alias{wide3m_to_monthly}
\title{Convertir formato wide (3m) a mensual}
\usage{
wide3m_to_monthly(df_wide, year_col = "Año", index_name = "ONI")
}
\arguments{
\item{df_wide}{data.frame en formato wide.}

\item{year_col}{nombre de la columna año.}

\item{index_name}{nombre del índice a asignar.}
}
\value{
data.frame mensual en formato tidy.
}
\description{
Convertir formato wide (3m) a mensual
}
